+++ midnight/BetterBlizzFrames.lua
@@ -5295,6 +5295,7 @@ PlayerEnteringWorld:SetScript("OnEvent", function()
     BBF.DarkmodeFrames()
     BBF.ClickthroughFrames()
     BBF.CheckForAuraBorders()
+    BBF.KeyDoublePress()
 end)
 PlayerEnteringWorld:RegisterEvent("PLAYER_ENTERING_WORLD")