+++ temp_cata/BetterBlizzFrames.lua
@@ -2914,6 +2914,7 @@ PlayerEnteringWorld:SetScript("OnEvent", function()
     if BetterBlizzFramesDB.playerFrameOCD then
         ChangeHotkeyWidth(32)
     end
+    BBF.KeyDoublePress()
 end)
 PlayerEnteringWorld:RegisterEvent("PLAYER_ENTERING_WORLD")